CertificateCredential uses AadClient
