+++ sdk/identity/azure-identity/azure/identity/_internal/__init__.py
@@ -34,6 +34,7 @@ from .aad_client import AadClient
 from .aad_client_base import AadClientBase
 from .auth_code_redirect_handler import AuthCodeRedirectServer
 from .aadclient_certificate import AadClientCertificate
+from .certificate_credential_base import CertificateCredentialBase
 from .exception_wrapper import wrap_exceptions
 from .msal_credentials import ConfidentialClientCredential, InteractiveCredential, PublicClientCredential
 from .msal_transport_adapter import MsalTransportAdapter, MsalTransportResponse
@@ -58,6 +59,7 @@ __all__ = [
     "AadClientBase",
     "AuthCodeRedirectServer",
     "AadClientCertificate",
+    "CertificateCredentialBase",
     "ConfidentialClientCredential",
     "get_default_authority",
     "InteractiveCredential",
