+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/certificate.py
@@ -5,8 +5,8 @@
 from typing import TYPE_CHECKING
 
 from .base import AsyncCredentialBase
-from .._authn_client import AsyncAuthnClient
-from ..._base import CertificateCredentialBase
+from .._internal import AadClient
+from ..._internal import CertificateCredentialBase
 
 if TYPE_CHECKING:
     from typing import Any
@@ -51,11 +51,10 @@ class CertificateCredential(CertificateCredentialBase, AsyncCredentialBase):
         if not scopes:
             raise ValueError("'get_token' requires at least one scope")
 
-        token = self._client.get_cached_token(scopes)
+        token = self._client.get_cached_access_token(scopes)
         if not token:
-            data = self._get_request_data(*scopes)
-            token = await self._client.request_token(scopes, form_data=data)
-        return token  # type: ignore
+            token = await self._client.obtain_token_by_client_certificate(scopes, self._certificate, **kwargs)
+        return token
 
-    def _get_auth_client(self, tenant_id, **kwargs):
-        return AsyncAuthnClient(tenant=tenant_id, **kwargs)
+    def _get_auth_client(self, tenant_id, client_id, **kwargs):
+        return AadClient(tenant_id, client_id, **kwargs)
